clean ups
- minimize hardcoded strings
- reduce info logs
- created shared.go for fns used across commands
