+++ release/internal/imagescanner/scanner.go
@@ -24,8 +24,6 @@ import (
 	"path/filepath"
 
 	"github.com/sirupsen/logrus"
-
-	"github.com/projectcalico/calico/release/internal/utils"
 )
 
 const (
@@ -60,11 +58,8 @@ func New(cfg Config) *Scanner {
 	}
 }
 
-// Scan sends a request to the image scanner to scan the given images.
-// The stream is the stream of the release.
-// The release flag indicates if the images are for a release which run image and code scans.
-// The outputDir is the directory to write the scan result to. If outputDir is empty, the scan result is not written to a file.
-func (i *Scanner) Scan(images []string, stream string, release bool, outputDir string) error {
+// Scan sends a request to the image scanner to scan the given images for the given product code and stream.
+func (i *Scanner) Scan(productCode, stream string, images []string, release bool, outputDir string) error {
 	var bucketPath, scanType string
 	if release {
 		scanType = "release"
@@ -92,7 +87,7 @@ func (i *Scanner) Scan(images []string, stream string, release bool, outputDir s
 	query := req.URL.Query()
 	query.Add("scan_type", scanType)
 	query.Add("scanner_select", i.config.Scanner)
-	query.Add("project_name", utils.ProductCode)
+	query.Add("project_name", productCode)
 	query.Add("project_version", stream)
 	req.URL.RawQuery = query.Encode()
 	logrus.WithFields(logrus.Fields{
